sched: Make nr_iowait_cpu() return 32-bit value
Runqueue ->nr_iowait counters are 32-bit anyway.

Propagate 32-bitness into other code, but don't try too hard.

Link: https://lore.kernel.org/r/20210422200228.1423391-3-adobriyan@gmail.com

+++ kernel/sched/core.c
@@ -4094,7 +4094,7 @@ unsigned int nr_iowait(void)
  * runnable.
  */
 
-unsigned long nr_iowait_cpu(int cpu)
+unsigned int nr_iowait_cpu(int cpu)
 {
 	struct rq *this = cpu_rq(cpu);
 	return atomic_read(&this->nr_iowait);
